requirements.txt
fastapi
uvicorn
python-dotenv
openai
pytest
pypdf
python-multipart
